chore: adds icons to deploy process (#2555)
* chore: adds icons to deploy process

* stuff

+++ scripts/setup-app.sh
@@ -5,11 +5,29 @@ fi
 
 API_URL="https://apollos-cluster-production.herokuapp.com/api"
 APP=$(curl -s --fail -H "x-api-key: $APOLLOS_API_KEY" -H "accept: text/plain" "$API_URL/config/$CHURCH/APP.NAME")
+ICON=$(curl -s --fail -H "x-api-key: $APOLLOS_API_KEY" -H "accept: text/plain" "$API_URL/config/$CHURCH/APP.ICON_URL")
+ICON_BG_COLOR=$(curl -s --fail -H "x-api-key: $APOLLOS_API_KEY" -H "accept: text/plain" "$API_URL/config/$CHURCH/APP.ICON_BG_COLOR")
+LOGO=$(curl -s --fail -H "x-api-key: $APOLLOS_API_KEY" -H "accept: text/plain" "$API_URL/config/$CHURCH/APP.LOGO_URL")
+WORDMARK=$(curl -s --fail -H "x-api-key: $APOLLOS_API_KEY" -H "accept: text/plain" "$API_URL/config/$CHURCH/APP.WORDMARK_URL")
 IOS_BUNDLE=$(curl -s --fail -H "x-api-key: $APOLLOS_API_KEY" -H "accept: text/plain" "$API_URL/config/$CHURCH/APP.APPLE_BUNDLE_ID")
 ANDROID_BUNDLE=$(curl -s --fail -H "x-api-key: $APOLLOS_API_KEY" -H "accept: text/plain" "$API_URL/config/$CHURCH/APP.ANDROID_PKG_ID")
 SERVER_URL=$(curl -s --fail -H "x-api-key: $APOLLOS_API_KEY" -H "accept: text/plain" "$API_URL/config/$CHURCH/APP.SERVER_URL")
 GOOGLE_MAPS_KEY=$(curl -s --fail -H "x-api-key: $APOLLOS_API_KEY" -H "accept: text/plain" "$API_URL/config/$CHURCH/APP_GOOGLE_MAPS_API_KEY")
 THEME=$(curl -s --fail -H "x-api-key: $APOLLOS_API_KEY" -H "accept: text/plain" "$API_URL/config/$CHURCH/APP.THEME")
+if [ -z "$APP" ] || [ -z "$ICON" ] || [ -z "$ICON_BG_COLOR" ] || [ -z "$LOGO" ] || [ -z "$WORDMARK" ] || [ -z "$IOS_BUNDLE" ] || [ -z "$ANDROID_BUNDLE" ] || [ -z "$SERVER_URL" ] || [ -z "$GOOGLE_MAPS_KEY" ] || [ -z "$THEME" ]; then
+    echo "Missing some variables:
+    APP.NAME=$APP
+    APP.ICON_URL=$ICON
+    APP.ICON_BG_COLOR=$ICON_BG_COLOR
+    APP.LOGO_URL=$LOGO
+    APP.WORDMARK_URL=$WORDMARK
+    APP.APPLE_BUNDLE_ID=$IOS_BUNDLE
+    APP.ANDROID_PKG_ID=$ANDROID_BUNDLE
+    APP.SERVER_URL=$SERVER_URL
+    APP_GOOGLE_MAPS_API_KEY=$GOOGLE_MAPS_KEY
+    APP.THEME=$THEME"
+    exit 1
+fi
 
 rm -rf deploys/mobile
 cp -r templates/mobile deploys/mobile
@@ -19,7 +37,7 @@ cd deploys/mobile || exit 1
 rm -rf fastlane
 mkdir fastlane
 
-printf $THEME > theme.json
+printf $THEME >theme.json
 
 CLEAN_APP=$(echo "$APP" | tr -d '[:space:]')
 npx react-native-rename "$CLEAN_APP" -b "$ANDROID_BUNDLE"
@@ -37,5 +55,16 @@ rm "ios/$CLEAN_APP.xcworkspace/xcshareddata/xcschemes/apolloschurchapp.xcscheme"
 echo "APP_DATA_URL=$SERVER_URL
 GOOGLE_MAPS_API_KEY=$GOOGLE_MAPS_KEY" >.env
 
+rm -rf node_modules
 yarn
 APP_DATA_URL=$SERVER_URL node scripts/get-introspection-data.js
+
+# graphics
+curl -s "$LOGO" --output src/img/logo.png
+curl -s "$WORDMARK" --output src/img/wordmark.png
+GRAPHICS_DIR=$(mktemp -d)
+node ../../scripts/generate-app-graphics.js "$GRAPHICS_DIR" "$ICON" "$ICON_BG_COLOR"
+cp "$GRAPHICS_DIR/rn-icon.png" src/img/icon.png
+yarn react-native set-splash --path "$GRAPHICS_DIR/splash.png" --background "$ICON_BG_COLOR"
+yarn react-native set-icon --path "$GRAPHICS_DIR/ios.png" --platform ios
+yarn react-native set-icon --path "$GRAPHICS_DIR/android.png" --platform android
